+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.2.0 - 2025-01-30
+### Added
+- Add UsageVal to replace Usage, and that allows receiving the value from the `--help` flag
+
+### Changed
+- Format the output of the default usage handler in nice columns
+- Allow the default usage handler to omit the CLI or ENV column based on the value of the `--help` flag
+
 ## 2.1.0 - 2022-09-12
 ### Changed
 - `Processor.Process` will now return an error when a non pointer type is used as an injection source